ZynAddSubFX/FLTK: updated to SVN revision 6970
Updated FLTK to SVN revision 6970 of branch-1.3:

    - Added Josef Vitu to CREDITS.
    - Moved OS X code base to the more moder Cocoa toolkit thanks to the awesome work of Manolo Gouy (STR #2221). This is a big one! I tested all test applications under 32-bit autoconf and Xcode, and a few apps under 64bit intel. No PPC testing was done. Please verify this patch if you have the machine!
    - Some fixes for the Cocoa port
    - Xcode project was not building Cocoa code!
    - Removed typedef that simply renamed char* to Fl_String, as discussed in the mailing list.
    - Checked in SebHoll's API mods, fixed indents.
        o Added user_data() to Fl_Tree_Item
        o Added insert() and add() methods that allow specification of Fl_Tree_Prefs
        o Changed Fl_Pixmap args to Fl_Image for more flexibility
        o Fixes for positioning of items in the presence of user icons
        o find_children() changed from protected -> public
    - Small fixed to SebHoll's user_data() mods (init + copy ctor),
    - Fumbeling about at the Cocoa implementation.
    - Fix for Mac's "ALT" key symbol: saucepan right side up, pancake not on floor.
    - Sorted Xcode source files (somewhat). Added Manolos patches to fix the Sys_Menu widget pointer.
    - Patced antialised fonts for Cocoa builds
    - Applied patch from dec 10th. It does not solve the problem that the About dialog drops behind the Sudoku window though...
    - Applied the 'recap.zip' files to the Cocoa patch. About diaog now correct. Much other stuff should work better now. Exciting!
    - New patches appliet for Cocoa port. Fixed(?) STR 2232 workaround for X11 keyrepeat bbbbuuuuuuggggg.
    - Added Manolo Gouy to CREDITS for his great work on porting FLTK 1.3 on Mac OS X to Cocoa. (Also fixed some space/tab inconsistencies).
    - Re-indented src/Fl_Input.cxx
    - Trying to fix update rectangles and dnd in Cocoa. Currently, invalidating a rectangle does not work. This can be seen when dragging data within an application, for example.

+++ plugins/zynaddsubfx/fltk/src/fl_line_style.cxx
@@ -40,7 +40,7 @@
 float fl_quartz_line_width_ = 1.0f;
 static enum CGLineCap fl_quartz_line_cap_ = kCGLineCapButt;
 static enum CGLineJoin fl_quartz_line_join_ = kCGLineJoinMiter;
-static float *fl_quartz_line_pattern = 0;
+static CGFloat *fl_quartz_line_pattern = 0;
 static int fl_quartz_line_pattern_size = 0;
 void fl_quartz_restore_line_style_() {
   CGContextSetLineWidth(fl_gc, fl_quartz_line_width_);
@@ -145,9 +145,9 @@ void fl_line_style(int style, int width, char* dashes) {
   fl_quartz_line_cap_ = Cap[(style>>8)&3];
   fl_quartz_line_join_ = Join[(style>>12)&3];
   char *d = dashes; 
-  static float pattern[16];
+  static CGFloat pattern[16];
   if (d && *d) {
-    float *p = pattern;
+	CGFloat *p = pattern;
     while (*d) { *p++ = (float)*d++; }
     fl_quartz_line_pattern = pattern;
     fl_quartz_line_pattern_size = d-dashes;
@@ -162,7 +162,7 @@ void fl_line_style(int style, int width, char* dashes) {
       dash = char(3*width);
       dot = gap = char(width);
     }
-    float *p = pattern;
+	CGFloat *p = pattern;
     switch (style & 0xff) {
     case FL_DASH:       *p++ = dash; *p++ = gap; break;
     case FL_DOT:        *p++ = dot; *p++ = gap; break;
@@ -172,7 +172,8 @@ void fl_line_style(int style, int width, char* dashes) {
     fl_quartz_line_pattern_size = p-pattern;
     fl_quartz_line_pattern = pattern;
   } else {
-    fl_quartz_line_pattern = 0; fl_quartz_line_pattern_size = 0;
+    fl_quartz_line_pattern = 0; 
+		fl_quartz_line_pattern_size = 0;
   }
   fl_quartz_restore_line_style_();
 #else
